requirements.txt
flask
flask-restful
flask-cors
requests
pandas
xlsxwriter
flask-login
ibm_db

python-dotenv
flask-restx
ibm-cloud-sdk-core==3.19.1

matplotlib